Move book routes
